Working flask site with Rekog functionality
