+++ Web Service Endpoint/templates/home.html
@@ -7,7 +7,7 @@
                     <a class="mr-2" href="#">{{ post.author }}</a>
                     <small class="text-muted">{{ post.date_posted }}</small>
                 </div> -->
-                <h2><a class="article-title">{{ post.title }}</a></h2>
+                <h2>Let's find your pill</h2>
                 <!-- <p class="article-content">{{ post.content }}</p> -->
                 
                 <form action="/home" method="POST" enctype="multipart/form-data">
