+++ Web Service Endpoint/templates/results.html
@@ -1,5 +1,5 @@
 {% extends "layout.html" %}
 {% block content %}
     <h1>Pill Results</h1>
-    <p>{{ response.json() }}</p>
+    <p>{{ result }}</p>
 {% endblock content %}
