--- Web Service Endpoint/templates/user.html
@@ -1,14 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-
-<h2>User: {{ title }}</h2>
-
-<p>{{ message }}</p>
-
-<ul>
-   {% for tweet in tweets %}
-   <li>{{ tweet.id }} - {{ tweet.text }}</li>
-   {% endfor %}
-</ul>
-
-{% endblock %}
